Goblins: fix dependency graph; fix typechecker error
There was strange inconsistent behavior between `stack build` and
`stack repl` where the former couldn't typecheck `goblins-test` but
the latter could.

This was solved by fixing a dependency cycle between `cs-ledger` and
`small-steps`, which results in a much more sane organization. I
glossed over this while trying to get things building, and am
surprised that `stack` didn't alert me to the cycle (whereas Marcin
tipped me off that `cabal new-build` error-ed on detecting the cycle.

I find it unintuitive that GHCi is able to typecheck with dependency
cycles where GHC isn't, but that must be a hidden idiosyncrasy in the
implementations.