goblins-test: allow for multiple runs & genome saving
The executable will now take an argument indicating the desired
number of runs. It will save the genomes of both failing and
successful goblins in a directory structure.

A command such as `ls -1 runs/**/pass | sort | uniq -c` can be
used to see the counts of the goblins deeped successfuly for each
PredicateFailure.