Goblins: add UPIREG & UPIVOTES `PredicateFailure`s
This required the addition of a number of `Goblin` instances, most
of which were generated via GHC Generics (implemented in the `goblins`
package, for which the stack.yaml git revision was bumped).

Training is successful on a significant number of these, though it takes
much longer with them added (likely becase we are testing "high up" at
the "plural" rules - e.g. UPIVOTES processes `[Vote]` as its Signal, so
the generator state space is much larger than for a `Vote` Signal). In
my last run, I saw 23 successfully bred goblins, out of 33 total
`PredicateFailure`s.

I had to increase the genome size because the state space is larger.