Goblins: match on PredicateFailure constructors only
Use Data.Data to only check if a goblin generated the right
PredicateFailure constructor. We don't compare equality on the fields
of the constructor.

We're only interested in triggering the right constructor - fields are
usually informational, and requiring them to be equal is overly
constraining the objective function.

With this change, we generate good goblins for 30 of the 33
`PredicateFailure` currently under test.